Fix background color of footer
Due to changes I made to fix validation issues, I "lost" the background
color in the footer. To fix this, I made the body background the
correct color, adjusted the "who" section to have a white background
and added some padding to the footer (by renaming the footer css
element) .container.footer to point to the correct element.